[DataLayer] Prepare beta version of the Adobe Client Data Layer
- moves itemUtils -> utils
