+++ src/scripts/DataLayerUtils.js
@@ -48,6 +48,78 @@ const utils = {
    */
   isObject: function(obj) {
     return (obj && typeof obj === 'object' && !Array.isArray(obj));
+  },
+  /**
+   * Determines whether the passed item configuration is a data configuration.
+   *
+   * @param {ItemConfig} itemConfig The item configuration.
+   * @returns {Boolean} true if the item configuration is a data configuration, false otherwise.
+   * @static
+   */
+  isDataConfig: function(itemConfig) {
+    if (!itemConfig) {
+      return false;
+    }
+    return (Object.keys(itemConfig).length === 1 && itemConfig.data);
+  },
+  /**
+   * Determines whether the passed item configuration is an event configuration.
+   *
+   * @param {ItemConfig} itemConfig The item configuration.
+   * @returns {Boolean} true if the item configuration is an event configuration, false otherwise.
+   * @static
+   */
+  isEventConfig: function(itemConfig) {
+    if (!itemConfig) {
+      return false;
+    }
+    const keys = Object.keys(itemConfig);
+    return (keys.length === 1 && itemConfig.event) ||
+      (keys.length === 2 && itemConfig.event && (itemConfig.info || itemConfig.data)) ||
+      (keys.length === 3 && itemConfig.event && itemConfig.info && itemConfig.data);
+  },
+  /**
+   * Determines whether the passed item is a listener configuration.
+   *
+   * @param {ItemConfig} itemConfig The item configuration.
+   * @returns {Boolean} true if the item is a listener on or listener off configuration, false otherwise.
+   * @static
+   */
+  isListenerConfig: function(itemConfig) {
+    return this.isListenerOnConfig(itemConfig) || this.isListenerOffConfig(itemConfig);
+  },
+  /**
+   * Determines whether the passed item configuration is a listener on configuration.
+   *
+   * @param {ItemConfig} itemConfig The item configuration.
+   * @returns {Boolean} true if the item configuration is a listener on configuration, false otherwise.
+   * @static
+   */
+  isListenerOnConfig: function(itemConfig) {
+    if (!itemConfig) {
+      return false;
+    }
+    const keys = Object.keys(itemConfig);
+    return (keys.length === 2 && itemConfig.on && itemConfig.handler) ||
+      (keys.length === 3 && itemConfig.on && itemConfig.handler && (itemConfig.scope || itemConfig.selector)) ||
+      (keys.length === 4 && itemConfig.on && itemConfig.handler && itemConfig.scope && itemConfig.selector);
+  },
+  /**
+   * Determines whether the passed item configuration is a listener off configuration.
+   *
+   * @param {ItemConfig} itemConfig The item configuration.
+   * @returns {Boolean} true if the item configuration is a listener off configuration, false otherwise.
+   * @static
+   */
+  isListenerOffConfig: function(itemConfig) {
+    if (!itemConfig) {
+      return false;
+    }
+    const keys = Object.keys(itemConfig);
+    return (keys.length === 1 && itemConfig.off) ||
+      (keys.length === 2 && itemConfig.off && itemConfig.handler) ||
+      (keys.length === 3 && itemConfig.off && itemConfig.handler && (itemConfig.scope || itemConfig.selector)) ||
+      (keys.length === 4 && itemConfig.off && itemConfig.handler && itemConfig.scope && itemConfig.selector);
   }
 };
